Added new schema template file
